+++ connect-inject/container_init_test.go
@@ -1,6 +1,7 @@
 package connectinject
 
 import (
+	"fmt"
 	"strings"
 	"testing"
 
@@ -20,11 +21,10 @@ func TestHandlerContainerInit(t *testing.T) {
 
 			Spec: corev1.PodSpec{
 				Containers: []corev1.Container{
-					corev1.Container{
+					{
 						Name: "web",
 					},
-
-					corev1.Container{
+					{
 						Name: "web-side",
 					},
 				},
@@ -38,14 +38,64 @@ func TestHandlerContainerInit(t *testing.T) {
 		Cmd    string // Strings.Contains test
 		CmdNot string // Not contains
 	}{
+		// The first test checks the whole template. Subsequent tests check
+		// the parts that change.
 		{
-			"Only service",
+			"Only service, whole template",
 			func(pod *corev1.Pod) *corev1.Pod {
 				pod.Annotations[annotationService] = "web"
 				return pod
 			},
-			`alias_service = "web"`,
-			`upstreams`,
+			`/bin/sh -ec export CONSUL_HTTP_ADDR="${HOST_IP}:8500"
+export CONSUL_GRPC_ADDR="${HOST_IP}:8502"
+
+# Register the service. The HCL is stored in the volume so that
+# the preStop hook can access it to deregister the service.
+cat <<EOF >/consul/connect-inject/service.hcl
+services {
+  id   = "${POD_NAME}-web-sidecar-proxy"
+  name = "web-sidecar-proxy"
+  kind = "connect-proxy"
+  address = "${POD_IP}"
+  port = 20000
+
+  proxy {
+    destination_service_name = "web"
+    destination_service_id = "web"
+  }
+
+  checks {
+    name = "Proxy Public Listener"
+    tcp = "${POD_IP}:20000"
+    interval = "10s"
+    deregister_critical_service_after = "10m"
+  }
+
+  checks {
+    name = "Destination Alias"
+    alias_service = "web"
+  }
+}
+
+services {
+  id   = "${POD_NAME}-web"
+  name = "web"
+  address = "${POD_IP}"
+  port = 0
+}
+EOF
+
+/bin/consul services register \
+  /consul/connect-inject/service.hcl
+
+# Generate the envoy bootstrap code
+/bin/consul connect envoy \
+  -proxy-id="${POD_NAME}-web-sidecar-proxy" \
+  -bootstrap > /consul/connect-inject/envoy-bootstrap.yaml
+
+# Copy the Consul binary
+cp /bin/consul /consul/connect-inject/consul`,
+			"",
 		},
 
 		{
@@ -55,7 +105,39 @@ func TestHandlerContainerInit(t *testing.T) {
 				pod.Annotations[annotationPort] = "1234"
 				return pod
 			},
-			"local_service_port = 1234",
+			`services {
+  id   = "${POD_NAME}-web-sidecar-proxy"
+  name = "web-sidecar-proxy"
+  kind = "connect-proxy"
+  address = "${POD_IP}"
+  port = 20000
+
+  proxy {
+    destination_service_name = "web"
+    destination_service_id = "web"
+    local_service_address = "127.0.0.1"
+    local_service_port = 1234
+  }
+
+  checks {
+    name = "Proxy Public Listener"
+    tcp = "${POD_IP}:20000"
+    interval = "10s"
+    deregister_critical_service_after = "10m"
+  }
+
+  checks {
+    name = "Destination Alias"
+    alias_service = "web"
+  }
+}
+
+services {
+  id   = "${POD_NAME}-web"
+  name = "web"
+  address = "${POD_IP}"
+  port = 1234
+}`,
 			"",
 		},
 
@@ -66,7 +148,15 @@ func TestHandlerContainerInit(t *testing.T) {
 				pod.Annotations[annotationUpstreams] = "db:1234"
 				return pod
 			},
-			`destination_name = "db"`,
+			`proxy {
+    destination_service_name = "web"
+    destination_service_id = "web"
+    upstreams {
+      destination_type = "service" 
+      destination_name = "db"
+      local_bind_port = 1234
+    }
+  }`,
 			"",
 		},
 
@@ -77,7 +167,16 @@ func TestHandlerContainerInit(t *testing.T) {
 				pod.Annotations[annotationUpstreams] = "db:1234:dc1"
 				return pod
 			},
-			`datacenter = "dc1"`,
+			`proxy {
+    destination_service_name = "web"
+    destination_service_id = "web"
+    upstreams {
+      destination_type = "service" 
+      destination_name = "db"
+      local_bind_port = 1234
+      datacenter = "dc1"
+    }
+  }`,
 			"",
 		},
 
@@ -92,58 +191,206 @@ func TestHandlerContainerInit(t *testing.T) {
 			`datacenter`,
 		},
 		{
-			"Check Destination Type Query Annotation",
+			"Upstream prepared query",
 			func(pod *corev1.Pod) *corev1.Pod {
 				pod.Annotations[annotationService] = "web"
 				pod.Annotations[annotationUpstreams] = "prepared_query:handle:1234"
 				return pod
 			},
-			`destination_type = "prepared_query"`,
-			`destination_type = "service"`,
-		},
-
-		{
-			"Check Destination Name Query Annotation",
-			func(pod *corev1.Pod) *corev1.Pod {
-				pod.Annotations[annotationService] = "web"
-				pod.Annotations[annotationUpstreams] = "prepared_query:handle:1234"
-				return pod
-			},
-			`destination_name = "handle"`,
+			`proxy {
+    destination_service_name = "web"
+    destination_service_id = "web"
+    upstreams {
+      destination_type = "prepared_query" 
+      destination_name = "handle"
+      local_bind_port = 1234
+    }
+  }`,
 			"",
 		},
 
 		{
-			"Service ID set to POD_NAME env var",
+			"Single Tag specified",
 			func(pod *corev1.Pod) *corev1.Pod {
 				pod.Annotations[annotationService] = "web"
-				pod.Annotations[annotationUpstreams] = "db:1234"
+				pod.Annotations[annotationPort] = "1234"
+				pod.Annotations[annotationTags] = "abc"
 				return pod
 			},
-			`id   = "${POD_NAME}-web-sidecar-proxy"`,
+			`services {
+  id   = "${POD_NAME}-web-sidecar-proxy"
+  name = "web-sidecar-proxy"
+  kind = "connect-proxy"
+  address = "${POD_IP}"
+  port = 20000
+  tags = ["abc"]
+
+  proxy {
+    destination_service_name = "web"
+    destination_service_id = "web"
+    local_service_address = "127.0.0.1"
+    local_service_port = 1234
+  }
+
+  checks {
+    name = "Proxy Public Listener"
+    tcp = "${POD_IP}:20000"
+    interval = "10s"
+    deregister_critical_service_after = "10m"
+  }
+
+  checks {
+    name = "Destination Alias"
+    alias_service = "web"
+  }
+}
+
+services {
+  id   = "${POD_NAME}-web"
+  name = "web"
+  address = "${POD_IP}"
+  port = 1234
+  tags = ["abc"]
+}`,
 			"",
 		},
 
 		{
-			"Proxy ID set to POD_NAME env var",
+			"Multiple Tags specified",
 			func(pod *corev1.Pod) *corev1.Pod {
 				pod.Annotations[annotationService] = "web"
-				pod.Annotations[annotationUpstreams] = "db:1234"
-				return pod
-			},
-			`-proxy-id="${POD_NAME}-web-sidecar-proxy"`,
-			"",
-		},
-
-		{
-			"Tags specified",
-			func(pod *corev1.Pod) *corev1.Pod {
-				pod.Annotations[annotationService] = "web"
-				pod.Annotations[annotationUpstreams] = "db:1234:dc1"
+				pod.Annotations[annotationPort] = "1234"
 				pod.Annotations[annotationTags] = "abc,123"
 				return pod
 			},
-			`tags = ["abc","123"]`,
+			`services {
+  id   = "${POD_NAME}-web-sidecar-proxy"
+  name = "web-sidecar-proxy"
+  kind = "connect-proxy"
+  address = "${POD_IP}"
+  port = 20000
+  tags = ["abc","123"]
+
+  proxy {
+    destination_service_name = "web"
+    destination_service_id = "web"
+    local_service_address = "127.0.0.1"
+    local_service_port = 1234
+  }
+
+  checks {
+    name = "Proxy Public Listener"
+    tcp = "${POD_IP}:20000"
+    interval = "10s"
+    deregister_critical_service_after = "10m"
+  }
+
+  checks {
+    name = "Destination Alias"
+    alias_service = "web"
+  }
+}
+
+services {
+  id   = "${POD_NAME}-web"
+  name = "web"
+  address = "${POD_IP}"
+  port = 1234
+  tags = ["abc","123"]
+}`,
+			"",
+		},
+
+		{
+			"Tags using old annotation",
+			func(pod *corev1.Pod) *corev1.Pod {
+				pod.Annotations[annotationService] = "web"
+				pod.Annotations[annotationPort] = "1234"
+				pod.Annotations[annotationConnectTags] = "abc,123"
+				return pod
+			},
+			`services {
+  id   = "${POD_NAME}-web-sidecar-proxy"
+  name = "web-sidecar-proxy"
+  kind = "connect-proxy"
+  address = "${POD_IP}"
+  port = 20000
+  tags = ["abc","123"]
+
+  proxy {
+    destination_service_name = "web"
+    destination_service_id = "web"
+    local_service_address = "127.0.0.1"
+    local_service_port = 1234
+  }
+
+  checks {
+    name = "Proxy Public Listener"
+    tcp = "${POD_IP}:20000"
+    interval = "10s"
+    deregister_critical_service_after = "10m"
+  }
+
+  checks {
+    name = "Destination Alias"
+    alias_service = "web"
+  }
+}
+
+services {
+  id   = "${POD_NAME}-web"
+  name = "web"
+  address = "${POD_IP}"
+  port = 1234
+  tags = ["abc","123"]
+}`,
+			"",
+		},
+
+		{
+			"Tags using old and new annotations",
+			func(pod *corev1.Pod) *corev1.Pod {
+				pod.Annotations[annotationService] = "web"
+				pod.Annotations[annotationPort] = "1234"
+				pod.Annotations[annotationTags] = "abc,123"
+				pod.Annotations[annotationConnectTags] = "abc,123,def,456"
+				return pod
+			},
+			`services {
+  id   = "${POD_NAME}-web-sidecar-proxy"
+  name = "web-sidecar-proxy"
+  kind = "connect-proxy"
+  address = "${POD_IP}"
+  port = 20000
+  tags = ["abc","123","abc","123","def","456"]
+
+  proxy {
+    destination_service_name = "web"
+    destination_service_id = "web"
+    local_service_address = "127.0.0.1"
+    local_service_port = 1234
+  }
+
+  checks {
+    name = "Proxy Public Listener"
+    tcp = "${POD_IP}:20000"
+    interval = "10s"
+    deregister_critical_service_after = "10m"
+  }
+
+  checks {
+    name = "Destination Alias"
+    alias_service = "web"
+  }
+}
+
+services {
+  id   = "${POD_NAME}-web"
+  name = "web"
+  address = "${POD_IP}"
+  port = 1234
+  tags = ["abc","123","abc","123","def","456"]
+}`,
 			"",
 		},
 
@@ -155,6 +402,78 @@ func TestHandlerContainerInit(t *testing.T) {
 			},
 			"",
 			`tags`,
+		},
+		{
+			"Metadata specified",
+			func(pod *corev1.Pod) *corev1.Pod {
+				pod.Annotations[annotationService] = "web"
+				pod.Annotations[annotationPort] = "1234"
+				pod.Annotations[fmt.Sprintf("%sname", annotationMeta)] = "abc"
+				pod.Annotations[fmt.Sprintf("%sversion", annotationMeta)] = "2"
+				return pod
+			},
+			`services {
+  id   = "${POD_NAME}-web-sidecar-proxy"
+  name = "web-sidecar-proxy"
+  kind = "connect-proxy"
+  address = "${POD_IP}"
+  port = 20000
+  meta = {
+    name = "abc"
+    version = "2"
+  }
+
+  proxy {
+    destination_service_name = "web"
+    destination_service_id = "web"
+    local_service_address = "127.0.0.1"
+    local_service_port = 1234
+  }
+
+  checks {
+    name = "Proxy Public Listener"
+    tcp = "${POD_IP}:20000"
+    interval = "10s"
+    deregister_critical_service_after = "10m"
+  }
+
+  checks {
+    name = "Destination Alias"
+    alias_service = "web"
+  }
+}
+
+services {
+  id   = "${POD_NAME}-web"
+  name = "web"
+  address = "${POD_IP}"
+  port = 1234
+  meta = {
+    name = "abc"
+    version = "2"
+  }
+}`,
+			"",
+		},
+
+		{
+			"No Metadata specified",
+			func(pod *corev1.Pod) *corev1.Pod {
+				pod.Annotations[annotationService] = "web"
+				return pod
+			},
+			"",
+			`meta`,
+		},
+
+		{
+			"Central config",
+			func(pod *corev1.Pod) *corev1.Pod {
+				pod.Annotations[annotationService] = "web"
+				return pod
+			},
+			"",
+			`meta`,
 		},
 	}
 
@@ -172,4 +491,156 @@ func TestHandlerContainerInit(t *testing.T) {
 			}
 		})
 	}
+}
+
+func TestHandlerContainerInit_centralConfig(t *testing.T) {
+	require := require.New(t)
+	h := Handler{
+		CentralConfig:   true,
+		DefaultProtocol: "grpc",
+	}
+	pod := &corev1.Pod{
+		ObjectMeta: metav1.ObjectMeta{
+			Annotations: map[string]string{
+				annotationService: "foo",
+			},
+		},
+
+		Spec: corev1.PodSpec{
+			Containers: []corev1.Container{
+				{
+					Name: "web",
+				},
+			},
+		},
+	}
+	container, err := h.containerInit(pod)
+	require.NoError(err)
+	actual := strings.Join(container.Command, " ")
+	require.Contains(actual, `
+# Create the central config's service registration
+cat <<EOF >/consul/connect-inject/central-config.hcl
+kind = "service-defaults"
+name = "foo"
+protocol = "grpc"
+EOF
+/bin/consul config write -cas -modify-index 0 \
+  /consul/connect-inject/central-config.hcl || true
+
+/bin/consul services register \
+  /consul/connect-inject/service.hcl
+
+# Generate the envoy bootstrap code
+/bin/consul connect envoy \
+  -proxy-id="${POD_NAME}-foo-sidecar-proxy" \
+  -bootstrap > /consul/connect-inject/envoy-bootstrap.yaml
+
+# Copy the Consul binary
+cp /bin/consul /consul/connect-inject/consul`)
+}
+
+func TestHandlerContainerInit_authMethod(t *testing.T) {
+	require := require.New(t)
+	h := Handler{
+		AuthMethod: "release-name-consul-k8s-auth-method",
+	}
+	pod := &corev1.Pod{
+		ObjectMeta: metav1.ObjectMeta{
+			Annotations: map[string]string{
+				annotationService: "foo",
+			},
+		},
+
+		Spec: corev1.PodSpec{
+			Containers: []corev1.Container{
+				{
+					Name: "web",
+					VolumeMounts: []corev1.VolumeMount{
+						{
+							Name:      "default-token-podid",
+							ReadOnly:  true,
+							MountPath: "/var/run/secrets/kubernetes.io/serviceaccount",
+						},
+					},
+				},
+			},
+		},
+	}
+	container, err := h.containerInit(pod)
+	require.NoError(err)
+	actual := strings.Join(container.Command, " ")
+	require.Contains(actual, `
+/bin/consul login -method="release-name-consul-k8s-auth-method" \
+  -bearer-token-file="/var/run/secrets/kubernetes.io/serviceaccount/token" \
+  -token-sink-file="/consul/connect-inject/acl-token" \
+  -meta="pod=${POD_NAMESPACE}/${POD_NAME}"
+
+/bin/consul services register \
+  -token-file="/consul/connect-inject/acl-token" \
+  /consul/connect-inject/service.hcl
+
+# Generate the envoy bootstrap code
+/bin/consul connect envoy \
+  -proxy-id="${POD_NAME}-foo-sidecar-proxy" \
+  -token-file="/consul/connect-inject/acl-token" \
+  -bootstrap > /consul/connect-inject/envoy-bootstrap.yaml`)
+}
+
+func TestHandlerContainerInit_authMethodAndCentralConfig(t *testing.T) {
+	require := require.New(t)
+	h := Handler{
+		AuthMethod:      "release-name-consul-k8s-auth-method",
+		CentralConfig:   true,
+		DefaultProtocol: "grpc",
+	}
+	pod := &corev1.Pod{
+		ObjectMeta: metav1.ObjectMeta{
+			Annotations: map[string]string{
+				annotationService: "foo",
+			},
+		},
+
+		Spec: corev1.PodSpec{
+			Containers: []corev1.Container{
+				{
+					Name: "web",
+					VolumeMounts: []corev1.VolumeMount{
+						{
+							Name:      "default-token-podid",
+							ReadOnly:  true,
+							MountPath: "/var/run/secrets/kubernetes.io/serviceaccount",
+						},
+					},
+				},
+			},
+		},
+	}
+	container, err := h.containerInit(pod)
+	require.NoError(err)
+	actual := strings.Join(container.Command, " ")
+	require.Contains(actual, `
+# Create the central config's service registration
+cat <<EOF >/consul/connect-inject/central-config.hcl
+kind = "service-defaults"
+name = "foo"
+protocol = "grpc"
+EOF
+/bin/consul login -method="release-name-consul-k8s-auth-method" \
+  -bearer-token-file="/var/run/secrets/kubernetes.io/serviceaccount/token" \
+  -token-sink-file="/consul/connect-inject/acl-token" \
+  -meta="pod=${POD_NAMESPACE}/${POD_NAME}"
+/bin/consul config write -cas -modify-index 0 \
+  -token-file="/consul/connect-inject/acl-token" \
+  /consul/connect-inject/central-config.hcl || true
+
+/bin/consul services register \
+  -token-file="/consul/connect-inject/acl-token" \
+  /consul/connect-inject/service.hcl
+
+# Generate the envoy bootstrap code
+/bin/consul connect envoy \
+  -proxy-id="${POD_NAME}-foo-sidecar-proxy" \
+  -token-file="/consul/connect-inject/acl-token" \
+  -bootstrap > /consul/connect-inject/envoy-bootstrap.yaml
+`)
 }